services/document_management/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.1  
python-dotenv==1.0.1
langchain==0.1.13
langchain-openai==0.1.1
langchain-community==0.0.29
PyMuPDF==1.23.3
python-multipart==0.0.5
pillow
pytesseract
pdf2image